split into utils non-exportable
